+++ src/main/java/chess/domain/board/Square.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 
 public class Square {
