+++ src/main/java/chess/domain/board/SquareFactory.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import java.util.ArrayList;
 import java.util.List;
