+++ src/test/java/chess/domain/board/SquareTest.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess.domain.board;
 
 import org.junit.jupiter.api.Test;
